pyproject.toml
[project]
name = "bodega"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "google-cloud-storage>=3.10.1",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.3",
    "python-multipart>=0.0.28",
    "ruff>=0.15.12",
]
